requirements.txt
requests
schedule
python-dotenv
slack_sdk
urllib3